+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.19.4-1.0.0.beta.6] - Unreleased
+
+### Added missing recipe for tool crafting table
+### Added all tool set for ruby
+### Added aquarium trident. - Textures not added yet.
+### Added shields for all custom ores. - Textures not added yet.
+
 ## [1.19.4-1.0.0.beta.5] - 2023-5-3
 
 ### Fixed issue with json update url.